% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roy.R
\name{roy}
\alias{roy}
\title{roy}
\usage{
roy(mod, var, trt, alpha = 0.05)
}
\arguments{
\item{mod}{modelo}

\item{var}{variavel}

\item{trt}{tratamento}

\item{alpha}{nivel alfa}
}
\description{
teste roy e bose para comparacoes multiplas
}
