% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roy.vc.R
\name{roy.vc}
\alias{roy.vc}
\title{roy.vc}
\usage{
roy.vc(mod, escores, trat, nrep, alpha = 0.05)
}
\arguments{
\item{mod}{modelo}

\item{escores}{vetor com os escores}

\item{trat}{vetor com os tratamentos}

\item{nrep}{numero de repeticoes}

\item{alpha}{nivel alfa}
}
\description{
teste roy e bose para comparacoes multiplas na primeira variavel canonica
}
